requirements.txt
pydantic~=2.7.3
passlib~=1.7.4
alembic~=1.13.1
sqlalchemy~=2.0.30
fastapi~=0.111.0
pydantic_settings
uvicorn~=0.30.1
asyncpg
python-jose